// source/src/models/table-filtered.ts
/* tslint:disable */
export interface TableFiltered {
  header?: string;
  field?: string;
  input?: string;
  metodo?: string;
  data?: any[];
  pattern?: string;
  type?: string;
  subcampo?: string;
  idsubcampo?: string;
  disabled?: boolean;
  required?: boolean;
  allowFiltered?: boolean;
  class?: string;
  classPadre?: string;
  placeHolder?: string;
  dataType?: string;
  disableSort?: boolean;
  valorCampoFiltro?: boolean;
  campoFiltro?: string;
  campoPrint?: string;
  campoData?: string;
  labels?: any;
}